[tTest] ttest added
The standalone tool of the t-test was added. It still needs some
improvements (no higher order attack yet, but it works for order 1). Added a config file also for the t test.

Added a common directory for shared classes (such as the Input class).

Added a "root" Makefile that compiles all the sub-modules.

Added some configuration example files.

Signed-off-by: davide <davidebotti@hotmail.it>